requirements.txt
# Core
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg[binary]>=3.1.0
psycopg-pool>=3.1.0

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
itsdangerous>=2.2.0
cryptography>=42.0.0

# Google APIs
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.118.0

# AI Stack
langchain>=1.2.0
langchain-groq>=1.1.0
langchain-community>=0.4.0
langchain-postgres>=0.0.17
langchain-huggingface>=1.2.0
langchain-text-splitters>=1.1.0
langgraph>=1.0.0
langgraph-checkpoint>=4.0.0
langgraph-checkpoint-postgres>=3.0.0

# ML
sentence-transformers>=2.3.0
flashrank>=0.2.0

# Groq
groq>=0.4.0

# Docs
pymupdf>=1.23.0
langchain-pymupdf4llm>=0.5.0
python-docx>=1.1.2
python-pptx>=1.0.2

# Utils
backoff>=2.2.0
itsdangerous>=2.2.0
Pillow>=12.1.1
pymupdf_layout>=0.0.1